Solvers

ProbNumDiffEq.jl provides two solvers: the EK1 and the EK0. Both based on extended Kalman filtering and smoothing, but the latter relies on evaluating the Jacobian of the vector field. For the best results, use the EK1.

All solvers are compatible with DAEs in mass-matrix ODE form. They also specialize on second-order ODEs: If the problem is of type SecondOrderODEProblem, it solves the second-order problem directly; this is more efficient than solving the transformed first-order problem and provides more meaningful posteriors [1].

API

ProbNumDiffEq.EK1 — Type

EK1(; order=3,
      smooth=true,
      prior=IWP(order),
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit(),
      kwargs...)

Gaussian ODE filter with first-order vector field linearization.

Arguments

order::Integer: Order of the integrated Wiener process (IWP) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

prior::AbstractODEFilterPrior: Prior to be used by the ODE filter. By default, uses a 3-times integrated Wiener process prior IWP(3). See also: Priors.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

Some additional kwargs relating to implicit solvers are supported; check out DifferentialEquations.jl's Extra Options page. Right now, we support autodiff, chunk_size, and diff_type. In particular, autodiff=false can come in handy to use finite differences instead of ForwardDiff.jl to compute Jacobians.

Examples

julia> solve(prob, EK1())

References

ProbNumDiffEq.EK0 — Type

EK0(; order=3,
      smooth=true,
      prior=IWP(order),
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit())

Gaussian ODE filter with zeroth-order vector field linearization.

Arguments

order::Integer: Order of the integrated Wiener process (IWP) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

prior::AbstractODEFilterPrior: Prior to be used by the ODE filter. By default, uses a 3-times integrated Wiener process prior IWP(3). See also: Priors.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

Examples

julia> solve(prob, EK0())

References

Probabilistic Exponential Integrators

ProbNumDiffEq.ExpEK — Function

ExpEK(; L, order=3, kwargs...)

Probabilistic exponential integrator

Probabilistic exponential integrators are a class of integrators for semi-linear stiff ODEs that provide improved stability by essentially solving the linear part of the ODE exactly. In probabilistic numerics, this amounts to including the linear part into the prior model of the solver.

ExpEK is therefore just a short-hand for EK0 with IOUP prior:

ExpEK(; order=3, L, kwargs...) = EK0(; prior=IOUP(order, L), kwargs...)

See also RosenbrockExpEK, EK0, EK1.

Arguments

See EK0 for available keyword arguments.

Examples

julia> prob = ODEProblem((du, u, p, t) -> (@. du = - u + sin(u)), [1.0], (0.0, 10.0))
julia> solve(prob, ExpEK(L=-1))

Reference

[2] Bosch et al, "Probabilistic Exponential Integrators", arXiv (2021)

ProbNumDiffEq.RosenbrockExpEK — Function

RosenbrockExpEK(; order=3, kwargs...)

Probabilistic Rosenbrock-type exponential integrator

A probabilistic exponential integrator similar to ExpEK, but with automatic linearization along the mean numerical solution. This brings the advantage that the linearity does not need to be specified manually, and the more accurate local linearization can sometimes also improve stability; but since the "prior" is adjusted at each step the probabilistic interpretation becomes more complicated.

RosenbrockExpEK is just a short-hand for EK1 with appropriete IOUP prior:

RosenbrockExpEK(; order=3, kwargs...) = EK1(; prior=IOUP(order, update_rate_parameter=true), kwargs...)

See also ExpEK, EK0, EK1.

Arguments

See EK1 for available keyword arguments.

Examples

julia> prob = ODEProblem((du, u, p, t) -> (@. du = - u + sin(u)), [1.0], (0.0, 10.0))
julia> solve(prob, RosenbrockExpEK())

Reference

[2] Bosch et al, "Probabilistic Exponential Integrators", arXiv (2021)

References

[2]

N. Bosch, P. Hennig and F. Tronarp. Probabilistic Exponential Integrators. arXiv:2305.14978 [math.NA] (2023).