Internals

ODEFilters.EKF0 — Type

EKF0(; prior=:ibm, order=1, diffusionmodel=:dynamic, smooth=true)

Gaussian ODE filtering with zeroth order extended Kalman filter.

Currently, only the integrated Brownian motion prior :ibm is supported. For the diffusionmodel, chose one of [:dynamic, :dynamicMV, :fixed, :fixedMV, :fixedMAP].

See also: EKF1

References:

M. Schober, S. Särkkä, and P. Hennig: A Probabilistic Model for the Numerical Solution of Initial Value Problems

F. Tronarp, H. Kersting, S. Särkkä, and P. Hennig: Probabilistic Solutions To Ordinary Differential Equations As Non-Linear Bayesian Filtering: A New Perspective

ODEFilters.EKF1 — Type

EKF1(; prior=:ibm, order=1, diffusionmodel=:dynamic, smooth=true)

Gaussian ODE filtering with first order extended Kalman filter

Currently, only the integrated Brownian motion prior :ibm is supported. For the diffusionmodel, chose one of [:dynamic, :dynamicMV, :fixed, :fixedMV, :fixedMAP].

See also: EKF0

References:

F. Tronarp, H. Kersting, S. Särkkä, and P. Hennig: Probabilistic Solutions To Ordinary Differential Equations As Non-Linear Bayesian Filtering: A New Perspective

ODEFilters.IEKS — Method

IEKS(; prior=:ibm, order=1, diffusionmodel=:dynamic, linearize_at=nothing)

Gaussian ODE filtering with iterated extended Kalman smoothing. To use it, use solve_ieks(prob, IEKS(), args...) instead of solve(prob, IEKS(), args...), since it is implemented as an outer loop around the solver.

Currently, only the integrated Brownian motion prior :ibm is supported. For the diffusionmodel, chose one of [:dynamic, :dynamicMV, :fixed, :fixedMV, :fixedMAP]. Just like the EKF1 it requires that the Jacobian of the rhs function is available.

See also: EKF0, EKF1, solve_ieks

References:

F. Tronarp, S. Särkkä, and P. Hennig: Bayesian ODE Solvers: The Maximum A Posteriori Estimate

ODEFilters.MAPFixedDiffusion — Type

Maximum a-posteriori Diffusion estimate when using an InverseGamma(1/2,1/2) prior

The mode of an InverseGamma(α,β) distribution is given by β/(α+1) To compute this in an on-line basis from the previous Diffusion, we reverse the computation to get the previous sum of residuals from Diffusion, and then modify that sum and compute the new Diffusion.

ODEFilters._rand — Function

Helper function to sample from our covariances, which often have a "cross" of zeros For the 0-cov entries the outcome of the sampling is deterministic!

ODEFilters.ibm — Function

Generate the discrete dynamics for a q-IBM model. INCLUDES AUTOMATIC PRECONDITIONING!

Careful: Dimensions are ordered differently than in probnum!

ODEFilters.predict! — Method

predict!(x_out, x_curr, Ah, Qh)

PREDICT step in Kalman filtering for linear dynamics models. In-place implementation of predict, saving the result in x_out.

\[m_{n+1}^P = A(h)*m_n P_{n+1}^P = A(h)*P_n*A(h) + Q(h)\]

See also: predict

ODEFilters.predict — Method

predict(x_curr::Gaussian, Ah::AbstractMatrix, Qh::AbstractMatrix)

PREDICT step in Kalman filtering for linear dynamics models.

\[m_{n+1}^P = A(h)*m_n P_{n+1}^P = A(h)*P_n*A(h) + Q(h)\]

See also: predict!

ODEFilters.remake_prob_with_jac — Method

remake_prob_with_jac(prob)

Add a jacobian function to the ODE function, using ModelingToolkit.jl.

ODEFilters.smooth — Method

smooth(x_curr::Gaussian, x_next_smoothed::Gaussian, Ah::AbstractMatrix, Qh::AbstractMatrix)

SMOOTH step of the (extended) Kalman smoother, or (extended) Rauch-Tung-Striebel smoother. It is implemented in Joseph Form:

\[m_{n+1}^P = A(h)*m_n P_{n+1}^P = A(h)*P_n*A(h) + Q(h) G = P_n * A(h)^T * (P_{n+1}^P)^{-1} m_n^S = m_n + G * (m_{n+1}^S - m_{n+1}^P) P_n^S = (I - G*A(h)) P_n (I - G*A(h))^T + G * Q(h) * G + G * P_{n+1}^S * G\]

ODEFilters.solve_ieks — Method

solve_ieks(prob::DiffEqBase.AbstractODEProblem, alg::IEKS, args...; iterations=10, kwargs...)

Solve method to be used with the IEKS. The IEKS works essentially by solving the ODE multiple times. solve_ieks therefore wraps a call to the standard solve method, passing args... and kwargs....

ODEFilters.update — Method

Vanilla UPDATE, without fancy checks or pre-allocation

Use this to test any "advanced" implementation against

ODEFilters.vanilla_ibm — Method

Same as above, but without the automatic preconditioning

OrdinaryDiffEq.perform_step! — Function

Perform a step

Not necessarily successful! For that, see step!(integ).

Basically consists of the following steps

Coordinate change / Predonditioning

Prediction step

Measurement: Evaluate f and Jf; Build z, S, H

Calibration; Adjust prediction / measurement covs if the diffusion model "dynamic"

Update step

Error estimation

Undo the coordinate change / Predonditioning