Solvers

ProbNumDiffEq.jl provides two solvers: the EK1 and the EK0. Both based on extended Kalman filtering and smoothing, but the latter relies on evaluating the Jacobian of the vector field. For the best results, use the EK1.

All solvers are compatible with DAEs in mass-matrix ODE form. They also specialize on second-order ODEs: If the problem is of type SecondOrderODEProblem, it solves the second-order problem directly; this is more efficient than solving the transformed first-order problem and provides more meaningful posteriors [1].

API

ProbNumDiffEq.EK1 — Type

EK1(; order=3,
      smooth=true,
      prior=IWP(order),
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit(),
      kwargs...)

Gaussian ODE filter with first-order vector field linearization.

Arguments

order::Integer: Order of the integrated Wiener process (IWP) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

prior::AbstractODEFilterPrior: Prior to be used by the ODE filter. By default, uses a 3-times integrated Wiener process prior IWP(3). See also: Priors.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

Some additional kwargs relating to implicit solvers are supported; check out DifferentialEquations.jl's Extra Options page. Right now, we support autodiff, chunk_size, and diff_type. In particular, autodiff=false can come in handy to use finite differences instead of ForwardDiff.jl to compute Jacobians.

Examples

julia> solve(prob, EK1())

References

ProbNumDiffEq.EK0 — Type

EK0(; order=3,
      smooth=true,
      prior=IWP(order),
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit())

Gaussian ODE filter with zeroth-order vector field linearization.

Arguments

order::Integer: Order of the integrated Wiener process (IWP) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

prior::AbstractODEFilterPrior: Prior to be used by the ODE filter. By default, uses a 3-times integrated Wiener process prior IWP(3). See also: Priors.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

Examples

julia> solve(prob, EK0())

References

References

[1] N. Bosch, F. Tronarp, P. Hennig: Pick-and-Mix Information Operators for Probabilistic ODE Solvers (2022) (link)