Solvers and Options

ProbNumDiffEq.jl provides mainly the following two solvers, both based on extended Kalman filtering and smoothing. For the best results we suggest using EK1, but note that it relies on the Jacobian of the vector field.

Solvers

All solvers are compatible with DAEs in mass-matrix ODE form, and specialize on second-order ODEs.

ProbNumDiffEq.EK1 — Type

EK1(; order=3, smooth=true,
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit(),
      kwargs...)

Gaussian ODE filter with first order vector field linearization.

Arguments

order::Integer: Order of the integrated Brownian motion (IBM) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

Some additional kwargs relating to implicit solvers are supported; check out DifferentialEquations.jl's Extra Options page. Right now, we support autodiff, chunk_size, and diff_type. In particular, autodiff=false can come in handy to use finite differences instead of ForwardDiff.jl to compute Jacobians.

References

ProbNumDiffEq.EK0 — Type

EK0(; order=3, smooth=true,
      diffusionmodel=DynamicDiffusion(),
      initialization=TaylorModeInit())

Gaussian ODE filter with zeroth order vector field linearization.

Arguments

order::Integer: Order of the integrated Brownian motion (IBM) prior.

smooth::Bool: Turn smoothing on/off; smoothing is required for dense output.

diffusionmodel::ProbNumDiffEq.AbstractDiffusion: See Diffusion models and calibration.

initialization::ProbNumDiffEq.InitializationScheme: See Initialization.

References

Diffusion models and calibration

In a nutshell: "Dynamic" diffusion models allow the diffusion to change in-between each solver step and are recommended in combination with adaptive step-sizes. "Fixed" diffusion models keep the diffusion constant and can be helpful in fixed-step settings or for debugging by reducing the complexity of the models.

For more information on the influence of diffusions, check out

N. Bosch, P. Hennig, F. Tronarp: Calibrated Adaptive Probabilistic ODE Solvers (2021)

ProbNumDiffEq.DynamicDiffusion — Type

DynamicDiffusion()

Recommended with adaptive steps

A local diffusion parameter is estimated at each step. Works well with adaptive steps.

ProbNumDiffEq.FixedDiffusion — Type

FixedDiffusion(; initial_diffusion=1.0, calibrate=true)

Recommended with fixed steps

If calibrate=true, the probabilistic solution is calibrated once at the end of the solve. An initial diffusion parameter can be passed, but this only has an effect if calibrate=false - which is typically not recommended.

ProbNumDiffEq.DynamicMVDiffusion — Type

DynamicMVDiffusion()

Only works with the EK0

A multi-variate version of DynamicDiffusion, where instead of a scalar a vector-valued diffusion is estimated. When using the EK0, this can be helpful when the scales of the different dimensions vary a lot.

ProbNumDiffEq.FixedMVDiffusion — Type

FixedMVDiffusion(; initial_diffusion=1.0, calibrate=true)

Only works with the EK0

A multi-variate version of FixedDiffusion, where instead of a scalar a vector-valued diffusion is estimated. When using the EK0, this can be helpful when the scales of the different dimensions vary a lot.

Initialization

ProbNumDiffEq.TaylorModeInit — Type

TaylorModeInit()

Recommended

Exact initialization via Taylor-mode automatic differentiation. Uses TaylorIntegration.jl. In case of errors, try ClassicSolverInit.

References:

N. Krämer, P. Hennig: Stable Implementation of Probabilistic ODE Solvers (2020)

ProbNumDiffEq.ClassicSolverInit — Type

ClassicSolverInit(; alg=OrdinaryDiffEq.Tsit5(), init_on_du=false)

Exact initialization with a classic ODE solver. The solver to be used can be set with the alg keyword argument. init_on_du specifies if ForwardDiff.jl should be used to compute the jacobian and initialize on the exact second derivative.

Not recommended for large solver orders, say order>4.