Solvers and Options

ProbNumDiffEq.jl provides mainly the following two solvers, both based on extended Kalman filtering and smoothing. For the best results we suggest using EK1, but note that it relies on the Jacobian of the vector field.

ProbNumDiffEq.EK1 — Type

EK1(; order=3, diffusionmodel=DynamicDiffusion(), smooth=true)

Gaussian ODE filtering with first order extended Kalman filtering.

All solvers use an integrated Brownian motion prior of order order. For the diffusionmodel, chose one of [DynamicDiffusion(), DynamicMVDiffusion(), FixedDiffusion(), FixedMVDiffusion()].

See also: EK0

References:

N. Bosch, P. Hennig, F. Tronarp: Calibrated Adaptive Probabilistic ODE Solvers (2021)

F. Tronarp, H. Kersting, S. Särkkä, and P. Hennig: Probabilistic Solutions To Ordinary Differential Equations As Non-Linear Bayesian Filtering: A New Perspective (2019)

ProbNumDiffEq.EK0 — Type

EK0(; order=3, diffusionmodel=DynamicDiffusion(), smooth=true)

Gaussian ODE filtering with zeroth order extended Kalman filtering.

All solvers use an integrated Brownian motion prior of order order. For the diffusionmodel, chose one of [DynamicDiffusion(), DynamicMVDiffusion(), FixedDiffusion(), FixedMVDiffusion()].

See also: EK1

References:

N. Bosch, P. Hennig, F. Tronarp: Calibrated Adaptive Probabilistic ODE Solvers (2021)

F. Tronarp, H. Kersting, S. Särkkä, and P. Hennig: Probabilistic Solutions To Ordinary Differential Equations As Non-Linear Bayesian Filtering: A New Perspective (2019)

M. Schober, S. Särkkä, and P. Hennig: A Probabilistic Model for the Numerical Solution of Initial Value Problems (2018)