Solvers and Options

EK0: Zeroth order extended Kalman filter and smoother

EK1: First order extended Kalman filter and smoother

IEKS: Iterated extended Kalman smoother

Missing docstring.

Missing docstring for EK0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EK1. Check Documenter's build log for details.

ProbNumDiffEq.IEKS — Type

IEKS(; prior=:ibm, order=1, diffusionmodel=:dynamic, linearize_at=nothing)

Gaussian ODE filtering with iterated extended Kalman smoothing. To use it, use solve_ieks(prob, IEKS(), args...) instead of solve(prob, IEKS(), args...), since it is implemented as an outer loop around the solver.

Currently, only the integrated Brownian motion prior :ibm is supported. For the diffusionmodel, chose one of [:dynamic, :dynamicMV, :fixed, :fixedMV, :fixedMAP]. Just like the EK1 it requires that the Jacobian of the rhs function is available.

See also: EK0, EK1, solve_ieks

References:

F. Tronarp, S. Särkkä, and P. Hennig: Bayesian ODE Solvers: The Maximum A Posteriori Estimate